refactor: Added a `CryptoKey` that unifies both `Salt` and `DerivedKey`.
Also, the `CryptoStore` is created from the stored salt and derived key. This allows to mutate both without having to write them if not needed.

+++ lib/model/crypto/salt.dart
@@ -1,18 +1,17 @@
 import 'dart:async';
 import 'dart:convert';
 
-import 'package:cipherlib/random.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:open_authenticator/model/crypto/key.dart';
 import 'package:open_authenticator/model/database/database.dart';
 import 'package:open_authenticator/model/totp/totp.dart';
 import 'package:simple_secure_storage/simple_secure_storage.dart';
 
 /// The salt provider instance.
-final saltProvider = AsyncNotifierProvider<StoredSalt, Salt?>(StoredSalt.new);
+final saltProvider = AsyncNotifierProvider<SaltNotifier, Salt?>(SaltNotifier.new);
 
 /// Allows to get and set the salt.
-class StoredSalt extends AsyncNotifier<Salt?> {
+class SaltNotifier extends AsyncNotifier<Salt?> {
   /// The password derived key storage key.
   static const String _kPasswordDerivedKeySaltKey = 'passwordDerivedKeySalt';
 
@@ -20,8 +19,8 @@ class StoredSalt extends AsyncNotifier<Salt?> {
   Future<Salt?> build() async {
     String? value = await SimpleSecureStorage.read(_kPasswordDerivedKeySaltKey);
     if (value != null) {
-      return Salt.fromRawValue(
-        value: base64.decode(value),
+      return Salt.base64Decode(
+        string: value,
       );
     }
     return await _tryRecoverSaltFromDatabase();
@@ -30,50 +29,70 @@ class StoredSalt extends AsyncNotifier<Salt?> {
   /// Tries to recover the salt from the database.
   Future<Salt?> _tryRecoverSaltFromDatabase() async {
     List<Totp> totps = await ref.read(appDatabaseProvider).listTotps();
-    Salt? salt = totps.firstOrNull?.encryptedData.encryptionSalt;
+    Map<Salt, int> salts = {};
+    (Salt, int)? best;
+    for (Totp totp in totps) {
+      int? count = salts[totp.encryptedData.encryptionSalt];
+      int newCount = count == null ? 1 : (count + 1);
+      salts[totp.encryptedData.encryptionSalt] = newCount;
+      if (best == null || newCount > best.$2) {
+        best = (totp.encryptedData.encryptionSalt, newCount);
+      }
+    }
+    Salt? salt = best?.$1;
     if (salt != null) {
       await _saveToLocalStorage(salt);
     }
     return salt;
   }
 
+  /// Generates a new salt if needed and updates the state.
+  Future<void> generateIfNeeded() async {
+    Salt? salt = await future;
+    if (salt != null) {
+      return;
+    }
+    salt = Salt.generate();
+    await changeSalt(salt);
+  }
+
   /// Changes the current salt.
-  Future<void> changeSalt(Salt salt) async {
-    _saveToLocalStorage(salt);
+  Future<void> changeSalt(Salt? salt) async {
+    await _saveToLocalStorage(salt);
     if (ref.mounted) {
       state = AsyncData(salt);
     }
   }
 
   /// Saves the given [salt] to the local storage.
-  Future<void> _saveToLocalStorage(Salt salt) async => await SimpleSecureStorage.write(_kPasswordDerivedKeySaltKey, base64.encode(salt.value));
-
-  /// Deletes the current salt from the local storage and resets the current state.
-  Future<void> deleteFromLocalStorage() async {
-    await SimpleSecureStorage.delete(_kPasswordDerivedKeySaltKey);
-    if (ref.mounted) {
-      state = const AsyncData(null);
+  Future<void> _saveToLocalStorage(Salt? salt) async {
+    if (salt == null) {
+      await SimpleSecureStorage.delete(_kPasswordDerivedKeySaltKey);
+    } else {
+      await SimpleSecureStorage.write(_kPasswordDerivedKeySaltKey, base64.encode(salt.value));
     }
   }
 }
 
 /// Represents a decoded salt.
-class Salt {
+class Salt extends CryptoKey {
   /// The salt length.
   static const int _saltLength = 256 ~/ 8;
 
-  /// The salt value.
-  final Uint8List value;
-
   /// Creates a new salt instance.
   const Salt.fromRawValue({
-    required this.value,
-  });
+    required super.value,
+  }) : super.fromRawValue();
+
+  /// Creates a new salt instance using the Base 64 encoded [string].
+  Salt.base64Decode({
+    required String string,
+  }) : this.fromRawValue(
+         value: base64.decode(string),
+       );
 
   /// Generates a random salt.
-  static Salt generate() => Salt.fromRawValue(
-    value: randomBytes(_saltLength),
-  );
+  Salt.generate() : super.generate(_saltLength);
 
   @override
   String toString() => base64.encode(value);
